+++ python/GafferImageTest/ImageWriterTest.py
@@ -1343,5 +1343,41 @@ class ImageWriterTest( GafferImageTest.ImageTestCase ) :
 		reader["refreshCount"].setValue( reader["refreshCount"].getValue() + 1 )
 		self.assertEqual( reader["out"].metadata()["openexr:dwaCompressionLevel"].value, 110.0 )
 
+	def testNameMetadata( self ) :
+
+		# Load an image with various layers.
+
+		reader = GafferImage.ImageReader()
+		reader["fileName"].setValue( "${GAFFER_ROOT}/python/GafferImageTest/images/multipart.exr" )
+		self.assertEqual(
+			set( reader["out"].channelNames() ),
+			{
+				"rgb.R", "rgb.G", "rgb.B",
+				"rgba.R", "rgba.G", "rgba.B", "rgba.A",
+				"depth.Z",
+			}
+		)
+
+		# Shuffle one of them into the primary RGBA layer.
+
+		shuffle = GafferImage.Shuffle()
+		shuffle["in"].setInput( reader["out"] )
+		shuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "R", "rgba.R" ) )
+		shuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "G", "rgba.G" ) )
+		shuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "B", "rgba.B" ) )
+		shuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "A", "rgba.A" ) )
+
+		# Write the image out and assert that it reads in again the same.
+
+		writer = GafferImage.ImageWriter()
+		writer["in"].setInput( shuffle["out"] )
+		writer["fileName"].setValue( os.path.join( self.temporaryDirectory(), "test.exr" ) )
+		writer["task"].execute()
+
+		rereader = GafferImage.ImageReader()
+		rereader["fileName"].setInput( writer["fileName"] )
+
+		self.assertImagesEqual( writer["in"], rereader["out"], ignoreMetadata = True, ignoreChannelNamesOrder = True )
+
 if __name__ == "__main__":
 	unittest.main()